Extract methods out of run file
