+++ cli/lib/app/index.js
@@ -1,4 +1,4 @@
-export { default as run } from "./run.js";
+export { default as run } from "./run/index.js";
 export { default as init } from "./init.js";
 export { default as clean } from "./clean.js";
 export { default as config } from "./config.js";
