--- cli/lib/app/run.js
@@ -1,143 +1,0 @@
-import glob from "glob";
-import path from "path";
-import chokidar from "chokidar";
-import figlet from "figlet";
-import * as print from "../helper/print.js";
-import { performLOD, copyOriginalArtifact } from "../lod.js";
-import { mergeOptionsWithConfigFile } from "../helper/config.js";
-import { MANIFEST_FILENAME } from "../constants.js";
-import {
-  createBaseFolderPathForFile,
-  getFilename,
-  getFolderPath,
-  getLastModified,
-  createFile,
-  getModelConfigFile,
-  getLodHashFile,
-  fileExists,
-  readFile,
-} from "../helper/files.js";
-import { startServer } from "../server/index.js";
-
-const readConfigFile = (file) => {
-  const configFile = readFile(file);
-  const parsedJson = JSON.parse(configFile);
-  return parsedJson;
-};
-
-const defaultRunOptions = {
-  source: "**/*.gltf",
-  config: "./lode-cli.config.json",
-  outputFoldername: "lode-build",
-};
-
-const generateHash = (originalFile, configuration) =>
-  getLastModified(originalFile) + JSON.stringify(configuration);
-
-const optimizeFile = ({ originalFile, levelDefinitions }) => {
-  const copiedOriginalFile = levelDefinitions[0].pathName;
-  const originalModified = () =>
-    getLastModified(copiedOriginalFile) < getLastModified(originalFile);
-
-  if (!fileExists(copiedOriginalFile) || originalModified()) {
-    copyOriginalArtifact(copiedOriginalFile, originalFile);
-  }
-  const levelsToBeReGenerated = levelDefinitions
-    .slice(1)
-    .filter((levelDefinition) => {
-      const hashFilePath = getLodHashFile(levelDefinition.pathName);
-      const newHash = generateHash(originalFile, levelDefinition.configuration);
-      return !(fileExists(hashFilePath) && readFile(hashFilePath) === newHash);
-    });
-  if (levelsToBeReGenerated.length > 0) {
-    performLOD({ originalFile, levelDefinitions: levelsToBeReGenerated });
-    levelsToBeReGenerated.forEach((levelDefinition) => {
-      const hashFilePath = getLodHashFile(levelDefinition.pathName);
-      const newHash = generateHash(originalFile, levelDefinition.configuration);
-      createFile(hashFilePath, newHash);
-    });
-    print.success("done");
-  }
-};
-
-const prepareFolders = (outputFoldername, sourceFiles) => {
-  return sourceFiles.reduce((agg, originalFile) => {
-    const filename = getFilename(originalFile);
-    const folderPath = getFolderPath(originalFile);
-    const modelConfig = readConfigFile(getModelConfigFile(originalFile));
-
-    const levelDefinitions = [];
-    for (let i = 0; i < modelConfig.levels.length; i++) {
-      const pathName = `./${outputFoldername}/${folderPath}/lod-${i}/${filename}`;
-      createBaseFolderPathForFile(pathName);
-      levelDefinitions.push({
-        pathName,
-        configuration: modelConfig.levels[i].configuration,
-      });
-    }
-
-    return { ...agg, [originalFile]: { originalFile, levelDefinitions } };
-  }, {});
-};
-
-const buildManifest = (outputFoldername, sourceFiles) => {
-  const manifest = sourceFiles.reduce((agg, sourceFile) => {
-    const modelConfig = readConfigFile(getModelConfigFile(sourceFile));
-    return {
-      ...agg,
-      [getFolderPath(sourceFile)]: modelConfig,
-    };
-  }, {});
-  createFile(
-    path.join(outputFoldername, MANIFEST_FILENAME),
-    JSON.stringify(manifest, null, 2)
-  );
-};
-
-const run = (commanderOptions) => {
-  print.warn(figlet.textSync("LODE", { horizontalLayout: "full" }));
-  const opts = mergeOptionsWithConfigFile(commanderOptions, defaultRunOptions);
-  const sourceFiles = glob.sync(opts.source);
-
-  if (sourceFiles.length === 0) {
-    print.error("No files found matching", opts.source);
-    print.error("Aborting");
-    return;
-  }
-  print.info("Preparing output folder:");
-  const fileStructure = prepareFolders(opts.outputFoldername, sourceFiles);
-  print.success("done");
-
-  print.info("Running initial LOD transformation:");
-  sourceFiles.forEach((file) => optimizeFile(fileStructure[file]));
-
-  buildManifest(opts.outputFoldername, sourceFiles);
-
-  if (opts.watch) {
-    print.info("watching files and running lode api server on port 3001...");
-
-    startServer({
-      outputFoldername: opts.outputFoldername,
-    });
-
-    chokidar
-      .watch(opts.source)
-      .on("add", (file) => optimizeFile(fileStructure[file]))
-      .on("change", (file) => optimizeFile(fileStructure[file]));
-    sourceFiles.forEach((srcFile) => {
-      const modelConfigFile = getModelConfigFile(srcFile);
-      chokidar
-        .watch(modelConfigFile, { awaitWriteFinish: true })
-        .on("change", () => {
-          const fileStructure = prepareFolders(opts.outputFoldername, [
-            srcFile,
-          ]);
-          optimizeFile(fileStructure[srcFile]);
-
-          buildManifest(opts.outputFoldername, sourceFiles);
-        });
-    });
-  }
-};
-
-export default run;
