+++ cli/lib/app/run/defaultRunOptions.js
@@ -1,0 +1,7 @@
+const defaultRunOptions = {
+  source: "**/*.gltf",
+  config: "./lode-cli.config.json",
+  outputFoldername: "lode-build",
+};
+
+export default defaultRunOptions;
