+++ cli/lib/app/run/generateHash.js
@@ -1,0 +1,6 @@
+import { getLastModified } from "../../helper/files.js";
+
+const generateHash = (originalFile, configuration) =>
+  getLastModified(originalFile) + JSON.stringify(configuration);
+
+export default generateHash;
