+++ cli/lib/app/run/readConfigFile.js
@@ -1,0 +1,8 @@
+import { readFile } from "../../helper/files.js";
+
+const readConfigFile = (file) => {
+  const configFile = readFile(file);
+  const parsedJson = JSON.parse(configFile);
+  return parsedJson;
+};
+export default readConfigFile;
